Refactor fetching of user data and update newUser logic
This commit refactors the code for fetching user data in the Layout component. It separates the fetching of the heliosUsername and avatarPath into separate useEffect hooks. The newUser state is now updated based on the fetched heliosUsername and avatarPath values. If both values are fetched and valid, newUser is set to false. Otherwise, it remains true. This improves the readability and maintainability of the code.